5_6_libft

+++ ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: hisasano <hisasano@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/25 12:47:05 by hisasano          #+#    #+#             */
-/*   Updated: 2025/05/05 13:48:15 by hisasano         ###   ########.fr       */
+/*   Updated: 2025/05/06 16:52:46 by hisasano         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 
 char	*ft_strdup(const char *s1)
 {
-	int		len;
-	int		i;
+	size_t	len;
+	size_t	i;
 	char	*dup;
 
 	len = ft_strlen(s1);
