+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: hisasano <hisasano@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/26 20:34:28 by hisasano          #+#    #+#             */
-/*   Updated: 2025/05/06 14:18:36 by hisasano         ###   ########.fr       */
+/*   Updated: 2025/05/06 16:42:49 by hisasano         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
 	int	len;
